move singletons folder setup to rppUtils
