+++ speed-dreams/source-2.2.3/src/simulatedDrivingAssistance/backendUnitTests/InterventionTests.cpp
@@ -15,27 +15,10 @@
 #define _SILENCE_EXPERIMENTAL_FILESYSTEM_DEPRECATION_WARNING 1
 #include <experimental/filesystem>
 
+#include "../rppUtils/RppUtils.hpp"
+
 #define INTERVENTION_TYPE_AMOUNT 5
 
-/// @brief Makes sure there is an empty singletons folder to be used by Singleton classes.
-///        Needs to be called once at the start of a method with any GetInstance() calls.
-inline void SetupSingletonsFolder()
-{
-    std::error_code errorCode;
-    std::experimental::filesystem::remove_all("Singletons", errorCode);
-    struct stat info;
-    char directory[256];
-    getcwd(directory, 256);
-    std::string workingDirecotory(directory);
-    workingDirecotory += "\\Singletons";
-    const char* wd = workingDirecotory.c_str();
-    int err = stat(wd, &info);
-    if (err != 0)
-    {
-        err = _mkdir(wd);
-        ASSERT_TRUE(err == 0);
-    }
-}
 
 InterventionType typesMediator[INTERVENTION_TYPE_AMOUNT] = { INTERVENTION_TYPE_NO_SIGNALS,
                                                              INTERVENTION_TYPE_ONLY_SIGNALS,
@@ -49,7 +32,8 @@ InterventionType typesMediator[INTERVENTION_TYPE_AMOUNT] = { INTERVENTION_TYPE_N
 
 TEST(MediatorTest, GetDistributedMediator)
 {
-    SetupSingletonsFolder();
+    bool succes = SetupSingletonsFolder();
+    ASSERT_TRUE(succes);
 
     SMediator* mediator1 = SMediator::GetInstance();
     SMediator* mediator2 = SMediator::GetInstance();
@@ -60,7 +44,8 @@ TEST(MediatorTest, GetDistributedMediator)
 /// @brief Tests if the Mediator sets and gets the interventionType correctly
 TEST(MediatorTest, GetIntervention)
 {
-    SetupSingletonsFolder();
+    bool succes = SetupSingletonsFolder();
+    ASSERT_TRUE(succes);
 
     SMediator* mediator = SMediator::GetInstance();
