+++ speed-dreams/source-2.2.3/src/simulatedDrivingAssistance/rppUtils/RppUtils.hpp
@@ -4,6 +4,9 @@
 #include "../../libs/portability/portability.h"
 #include <iostream>
 
+#define _SILENCE_EXPERIMENTAL_FILESYSTEM_DEPRECATION_WARNING 1
+#include <experimental/filesystem>
+
 inline float stringToFloat(std::string s)
 {
     try { return std::stof(s); }
@@ -49,4 +52,40 @@ inline bool FindFileDirectory(std::string& p_knownPathToFile, const std::string&
     }
 
     return false;
+}
+
+/// @brief  Makes sure there is an empty singletons folder to be used by Singleton classes.
+///         Needs to be called once at the start of a method with any GetInstance() calls.
+/// @return Boolean indicating whether the setup succeeded or not
+inline bool SetupSingletonsFolder()
+{
+    std::error_code errorCode;
+    std::experimental::filesystem::remove_all("Singletons", errorCode);
+    if (errorCode.value() != 0)
+    {
+        std::cerr << "Something went wrong when removing the Singleton folder: " << errorCode.value();
+        return false;
+    }
+
+    // set up singleton folder
+    struct stat info;
+    char directory[256];
+    getcwd(directory, 256);
+    std::string workingDirecotory(directory);
+    workingDirecotory += "\\Singletons";
+    const char* wd = workingDirecotory.c_str();
+    int err = stat(wd, &info);
+    if (err != 0 && err != -1)
+    {
+        std::cerr << "Could not delete Singletons folder" << std::endl;
+        return false;
+    }
+    err = _mkdir(wd);
+    if (err != 0)
+    {
+        std::cerr << "Could not create singletons folder" << std::endl;
+        return false;
+    }
+
+    return true;
 }
